# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Diet Assistant API
  description: |
    RESTful API for AI-powered diet planning and nutrition management system.
    
    ## Features
    - JWT-based authentication
    - Food inventory management
    - Meal tracking and recording
    - AI-powered meal plan generation
    - Nutrition analysis and statistics
    - Conversation with AI assistant
    
    ## Authentication
    Most endpoints require JWT Bearer token authentication.
    Include the token in the Authorization header: `Bearer {token}`
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:9090/api/v1
    description: Development server
  - url: https://api.yourdomain.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
    Error:
      type: object
      properties:
        code:
          type: integer
          example: 40001
        message:
          type: string
          example: "Invalid parameters"
        error:
          type: string
          example: "Detailed error message"
        timestamp:
          type: integer
          format: int64
          example: 1699999999
    SuccessResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: "success"
        timestamp:
          type: integer
          format: int64
          example: 1699999999
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        total_pages:
          type: integer
          example: 5
...
    Food:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        user_id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Chicken Breast"
        category:
          type: string
          enum: [meat, vegetable, fruit, grain, other]
          example: "meat"
        price:
          type: number
          format: float
          example: 12.99
        unit:
          type: string
          example: "g"
        protein:
          type: number
          format: float
          example: 31.0
        carbs:
          type: number
          format: float
          example: 0.0
        fat:
          type: number
          format: float
          example: 3.6
        fiber:
          type: number
          format: float
          example: 0.0
        calories:
          type: number
          format: float
          example: 165.0
        available:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /foods:
    get:
      tags:
        - Foods
      summary: List food items
      description: Get paginated list of food items with optional filtering
      operationId: listFoods
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
        - name: category
          in: query
          schema:
            type: string
            enum: [meat, vegetable, fruit, grain, other]
        - name: available
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of foods
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Food'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
    
    post:
      tags:
        - Foods
      summary: Create food item
      description: Add a new food item to inventory
      operationId: createFood
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - category
              properties:
                name:
                  type: string
                category:
                  type: string
                  enum: [meat, vegetable, fruit, grain, other]
                price:
                  type: number
                unit:
                  type: string
                protein:
                  type: number
                carbs:
                  type: number
                fat:
                  type: number
                fiber:
                  type: number
                calories:
                  type: number
                available:
                  type: boolean
      responses:
        '201':
          description: Food created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Food'
  /foods/{id}:
    get:
      tags:
        - Foods
      summary: Get food item
      description: Get food item by ID
      operationId: getFood
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Food item details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Food'
        '404':
          description: Food not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      tags:
        - Foods
      summary: Update food item
      description: Update existing food item
      operationId: updateFood
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                category:
                  type: string
                price:
                  type: number
                available:
                  type: boolean
      responses:
        '200':
          description: Food updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Food'
    
    delete:
      tags:
        - Foods
      summary: Delete food item
      description: Remove food item from inventory
      operationId: deleteFood
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Food deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...